Add support for SuperCard/3in1 Plus
* Add new error message for detecting if run on DSi/3DS consoles as this
program isn't really compatible with those consoles for obvious reasons.
* Add initial support for SuperCard Lite (and possibly other SuperCard
varients). Note that saves are not currently functional though.
* Add back initial support for EZ Flash 3 in 1 Plus. Note that NorFlash
commands appear to not be working at the moment. RAM mode untested. New
code added for detecting 3 in 1 Plus when detecting max allowed file
size. This should allow writing 64MB gba roms to 3 in 1 Plus...once
NorFlash stuff is fixed that is. :P
* Tested as working properly with regular EZ Flash 3 in 1 carts.
* Note that version 0.58 already exists but we do not have source code
for it. (source was released only for 0.57. Rudolph could not find the
last version's source code unfortunately. This is why 3 in 1 Plus
support is incomplete. (it at least detects it now which original 0.57
could not do and will in theory allow writing 64MB GBA roms once
NorFlash stuff is fixed for that cart)