Fix 3 in 1 Plus NorFlash
* NorFlash read/write now working for 3 in 1 Plus! Big thanks to
cory1492 for he had managed to make available the version of source code
of gbaldr that had the code needed to make this work. Also thanks to
stl25 for helping test 3in1 Plus support as I do not own a 3 in 1 Plus
myself.

* 64MB roms appear to write to NorFlash correctly with 3 in 1 Plus
however note that this does not mean retail roms that used that size
will work (not without patching anyways). 3 in 1 Plus had 64MB NorFlash
but doen't use normal page switch commands so retail roms will need
patching to use this.

* r4tf bool set to false for SuperCard and Rumble menu disabled for
SuperCard. Program would crash if user attempts to use R shoulder button
to access rumble menu if a SuperCard is inserted and a NDS file for soft
reset was available. This has been fixed by disabling this menu for
SuperCard. The rumble menu will not work as intended for SuperCards
anyways.
* Soft Reset for rumble menu is currently broken. (has been since
initial rebuild). The method the program uses for loading NDS files
likely needs to be rewritten.
* Hitting X after flashing a game to NorFlash on 3 in 1 Plus may cause
hang. The game flashed will still be bootable provided the user power
cycles the console and boots the card as normal. No plans to fix as I'm
unsure why it's happening and it's a minor issue. Hitting X after
switching to NorFlash without flashing anything seems to still work
though. Feel free to create a push request if you managed to fix this.
;)